--- pytorch_forecasting/layers/_mlstm/__init__.py
@@ -1,7 +1,0 @@
-"""mLSTM layer"""
-
-from pytorch_forecasting.layers._mlstm.cell import mLSTMCell
-from pytorch_forecasting.layers._mlstm.layer import mLSTMLayer
-from pytorch_forecasting.layers._mlstm.network import mLSTMNetwork
-
-__all__ = ["mLSTMCell", "mLSTMLayer", "mLSTMNetwork"]
