+++ pytorch_forecasting/layers/_recurrent/__init__.py
@@ -1,0 +1,21 @@
+"""Recurrent Layers for Pytorch-Forecasting"""
+
+from pytorch_forecasting.layers._recurrent._mlstm import (
+    mLSTMCell,
+    mLSTMLayer,
+    mLSTMNetwork,
+)
+from pytorch_forecasting.layers._recurrent._slstm import (
+    sLSTMCell,
+    sLSTMLayer,
+    sLSTMNetwork,
+)
+
+__all__ = [
+    "mLSTMCell",
+    "mLSTMLayer",
+    "mLSTMNetwork",
+    "sLSTMCell",
+    "sLSTMLayer",
+    "sLSTMNetwork",
+]
